src/tasks/task-5-deployment/Streamlit-app/requirements.txt
streamlit-option-menu
geopandas
millify
folium
streamlit_folium
pycaret==3.0.0rc2